requirements.txt
asgiref==3.8.1
billiard==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==2.10
kombu==5.3.7
packaging==24.1
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.4
vine==5.1.0
wcwidth==0.2.13
